# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: NextSmartShip openapi
  description: NextSmartShip openapi for oms and wms
  version: ""
host: 0.0.0.0:80
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /upload_image:
    post:
      tags:
      - file
      summary: upload_image file
      operationId: file#upload_image
      consumes:
      - multipart/form-data
      parameters:
      - name: upload_image_request_body
        in: formData
        required: true
        schema:
          $ref: '#/definitions/FileUploadImageRequestBody'
          required:
          - file
          - file_name
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/FileUploadImageResponseBody'
            required:
            - url
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/FileUploadImageUnauthorizedResponseBody'
      schemes:
      - http
definitions:
  FileUploadImageRequestBody:
    title: FileUploadImageRequestBody
    type: object
    properties:
      file:
        type: string
        description: file
        example: 1.jpg
        maxLength: 864000
      file_name:
        type: string
        description: file name
        example: xxx
        maxLength: 50
    example:
      file: 1.jpg
      file_name: xxx
    required:
    - file
    - file_name
  FileUploadImageResponseBody:
    title: FileUploadImageResponseBody
    type: object
    properties:
      url:
        type: string
        description: image URL
        example: http://images.example
        maxLength: 300
    example:
      url: http://images.example
    required:
    - url
  FileUploadImageUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: upload_image_Unauthorized_response_body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...